Add some tests, and Translate interface

+++ intrip/Model/Entities/ItemFixer.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 struct ItemFixer: Codable {
-    init(success: Bool,timestamp: Double,base: String, date : String, rates: Dictionary<String, Double> ){
+    init(success: Bool,timestamp: Int,base: String, date : String, rates: Dictionary<String, Double> ){
         self.success = success
         self.timestamp = timestamp
         self.base = base
@@ -16,7 +16,7 @@ struct ItemFixer: Codable {
         self.rates = rates
     }
     var success: Bool
-    var timestamp: Double
+    var timestamp: Int
     var base: String
     var date: String
     var rates: Dictionary<String, Double>
